+++ packages/dso-toolkit/src/components/input-number/input-number.models.ts
@@ -1,9 +1,0 @@
-export interface InputNumber {
-  label?: string;
-  id: string;
-  min?: number;
-  max?: number;
-  count: number;
-  minusButtonInactive: boolean;
-  plusButtonInactive: boolean;
-}