+++ packages/dso-toolkit/src/components/list-button/list-button.args.ts
@@ -4,12 +4,10 @@ import { HandlerFunction } from "@storybook/addon-actions";
 import { noControl } from "../../storybook/index.js";
 
 import { ListButton } from "./list-button.models.js";
-import { InputNumber } from "../input-number";
 
 export interface ListButtonArgs {
   label: string;
   hasInputNumber?: boolean;
-  inputNumber?: InputNumber;
   disabled?: boolean;
   sublabel?: string;
   subcontent?: string;
@@ -18,6 +16,8 @@ export interface ListButtonArgs {
   count?: number;
   min?: number;
   max?: number;
+  minusButtonInactive: boolean;
+  plusButtonInactive: boolean;
   manual: boolean;
   dsoCountChange: HandlerFunction;
   dsoSelectedChange: HandlerFunction;
@@ -73,6 +73,16 @@ export const listButtonArgTypes: ArgTypes<ListButtonArgs> = {
     },
   },
   manual: {
+    control: {
+      type: "boolean",
+    },
+  },
+  minusButtonInactive: {
+    control: {
+      type: "boolean",
+    },
+  },
+  plusButtonInactive: {
     control: {
       type: "boolean",
     },