+++ packages/dso-toolkit/src/components/list-button/list-button.models.ts
@@ -1,10 +1,6 @@
-import { InputNumber } from "../input-number/input-number.models.js";
-
 export interface ListButton {
   label: string;
-  /** Deprecated: used only in html/css component */
   hasInputNumber?: boolean;
-  inputNumber?: InputNumber;
   disabled?: boolean;
   sublabel?: string;
   subcontent?: string;
@@ -12,6 +8,8 @@ export interface ListButton {
   count?: number;
   min?: number;
   max?: number;
+  minusButtonInactive?: boolean;
+  plusButtonInactive?: boolean;
   checked?: boolean;
   manual?: boolean;
   dsoCountChange?: (e: CustomEvent<ListButtonChangeEvent>) => void;