+++ storybook/src/components/list-button/list-button.css-template.ts
@@ -4,11 +4,12 @@ import { html, nothing } from "lit-html";
 import { classMap } from "lit-html/directives/class-map.js";
 import { unsafeHTML } from "lit-html/directives/unsafe-html.js";
 import { ComponentImplementation } from "../../templates";
+import { ifDefined } from "lit-html/directives/if-defined.js";
 
 export const cssListButton: ComponentImplementation<ListButton> = {
   component: "listButton",
   implementation: "html-css",
-  template: ({ inputNumberTemplate, iconTemplate }) =>
+  template: ({ buttonTemplate, iconTemplate }) =>
     function listButtonTemplate(listButton) {
       function listButtonComponentTemplate({
         count,
@@ -40,11 +41,38 @@ export const cssListButton: ComponentImplementation<ListButton> = {
       }
 
       return html`
-        ${listButton.hasInputNumber && listButton.inputNumber
+        ${listButton.hasInputNumber
           ? html`
               <div class="dso-button-group">
                 ${listButtonComponentTemplate(listButton)}
-                ${inputNumberTemplate({ ...listButton.inputNumber, count: listButton.count ?? 0 })}
+                <div class="dso-input-number">
+                  ${buttonTemplate({
+                    type: "button",
+                    label: "Aantal verlagen",
+                    variant: "tertiary",
+                    disabled: listButton.minusButtonInactive,
+                    icon: { icon: "minus-square" },
+                    iconMode: "only",
+                  })}
+                  <input
+                    type="number"
+                    readonly
+                    tabindex="-1"
+                    min=${ifDefined(listButton.min)}
+                    max=${ifDefined(listButton.max)}
+                    class="dso-input-step-counter"
+                    aria-label="Aantal"
+                    value=${listButton.count}
+                  />
+                  ${buttonTemplate({
+                    type: "button",
+                    label: "Aantal verhogen",
+                    variant: "tertiary",
+                    disabled: listButton.plusButtonInactive,
+                    icon: { icon: "plus-square" },
+                    iconMode: "only",
+                  })}
+                </div>
               </div>
             `
           : html` ${listButtonComponentTemplate(listButton)} `}